move timer to utils dir
