+++ src/common/session/session-entity.js
@@ -4,7 +4,7 @@ import { stringify } from '../util/stringify'
 import { documentAddEventListener } from '../event-listener/event-listener-opts'
 import { ee } from '../event-emitter/contextual-ee'
 import { subscribeToVisibilityChange } from '../window/page-visibility'
-import { Timer } from '../timer/timer'
+import { Timer } from '../util/timer'
 import LocalStorage from '../storage/local-storage.js'
 import FPC from '../storage/first-party-cookies'
 import { getConfiguration } from '../config/config'
