+++ src/common/util/timer.js
